REF-1 Refactor: Move all action.types constants to another file
