+++ frontend/src/actions/index.js
@@ -14,35 +14,30 @@ import {
   editCommentAPI
 } from '../utils/api.js'
 
-// define action constants
-export const GET_ALL_POSTS = 'GET_ALL_POSTS'
-export const RECEIVE_POSTS = 'RECEIVE_POSTS'
-export const GET_POST = 'GET_POST'
-export const ADD_POST = 'ADD_POST'
-export const EDIT_POST = 'EDIT_POST'
-export const REMOVE_POST = 'REMOVE_POST'
-
-export const RECEIVE_CATEGORIES = 'RECEIVE_CATEGORIES'
-
-export const RECEIVE_COMMENTS = 'RECEIVE_COMMENTS'
-
-export const ADD_COMMENT = 'ADD_COMMENT'
-export const REMOVE_COMMENT = 'REMOVE_COMMENT'
-export const EDIT_COMMENT = 'EDIT_COMMENT'
-
-export const CHANGE_SORT_ORDER = 'CHANGE_SORT_ORDER'
-export const DESCENDING_ORDER = 'desc'
-export const ASCENDING_ORDER = 'asc'
-
-export const CHANGE_COMMENTS_ORDER = 'CHANGE_COMMENTS_ORDER'
-
-export const INCREASE_POST_SCORE = 'INCREASE_POST_SCORE'
-export const DECREASE_POST_SCORE = 'DECREASE_POST_SCORE'
-
-export const INCREASE_COMMENT_SCORE = 'INCREASE_COMMENT_SCORE'
-export const DECREASE_COMMENT_SCORE = 'DECREASE_COMMENT_SCORE'
-
-export const SET_CATEGORY_FILTER = 'SET_CATEGORY_FILTER'
+import {
+  RECEIVE_POSTS,
+  ADD_POST,
+  EDIT_POST,
+  REMOVE_POST,
+  
+  RECEIVE_CATEGORIES,
+  
+  RECEIVE_COMMENTS,
+  ADD_COMMENT,
+  REMOVE_COMMENT,
+  EDIT_COMMENT,
+  
+  CHANGE_SORT_ORDER,
+  CHANGE_COMMENTS_ORDER,
+  
+  INCREASE_POST_SCORE,
+  DECREASE_POST_SCORE,
+  
+  INCREASE_COMMENT_SCORE,
+  DECREASE_COMMENT_SCORE,
+  
+  SET_CATEGORY_FILTER
+} from './types'
 
 // define action creators
