+++ frontend/src/reducers/sortReducer.js
@@ -1,7 +1,7 @@
 import {
   CHANGE_SORT_ORDER,
   DESCENDING_ORDER
-} from '../actions'
+} from '../actions/types'
 
 const defaultSortOrder = {
   field: 'voteScore',
